Bundle API: remove BundleInfo from the status
The BundleInfo struct encodes assumptions of bundles that do not exist
in the Bundle API. The only field we currently populate in the
BundleInfo struct in the plain manifest provisioner is the objects
field, so we should remove (at least for now) the package, name and
version fields.

We should also remove the objects field because not all Bundle formats
will have a concept of bundle objects at bundle unpack time. For
example, a helm chart bundle type might only have templates at bundle
unpack time and only generate a helm release with objects for a bundle
instance.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>